+++ playwright/e2e/colors.spec.ts
@@ -1,0 +1,6 @@
+import { test } from "@playwright/test"
+import { baseTests } from "./baseTest"
+
+test.describe("Colors Page Visual Test", () => {
+  baseTests("colors")
+})