+++ playwright/e2e/icon-states.spec.ts
@@ -1,0 +1,6 @@
+import { test } from "@playwright/test"
+import { baseTests } from "./baseTest"
+
+test.describe("Icon States Page Visual Test", () => {
+  baseTests("icon-states", "Icon States")
+})